🙀 chore: 커스텀 select에서 사용하는 type 및 enum type 폴더로 분리
