+++ apps/client/src/shared/types/customSelectType.ts
@@ -1,0 +1,10 @@
+export enum SelectSize {
+  SMALL = 'SMALL',
+
+  MEDIUM = 'MEDIUM',
+}
+
+export type TOption = {
+  value: string;
+  label: string;
+};
